testing: finalize test coverage
Closes #123713
