+++ extensions/vscode-api-tests/package.json
@@ -45,7 +45,6 @@
     "terminalDataWriteEvent",
     "terminalDimensions",
     "tunnels",
-    "testCoverage",
     "testObserver",
     "textSearchProvider",
     "timeline",
