+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "code-oss-dev",
   "version": "1.88.0",
-  "distro": "7ca938298e57ad434ea8807e132707055458a749",
+  "distro": "ff3bff60edcc6e1f7269509e1673036c00fa62bd",
   "author": {
     "name": "Microsoft Corporation"
   },
